Controllability model was not sensitive enough
And was spitting that no variable was correlated despite there being a clear correlation. It was because the range of probabilities was very small. So I ended up using a lm with logit link instead of a glm. The species level figure, the text in the results and the table in the Suppporting Information have been updated